Move broadcaster to gateway/ dir
